Periodically check and update ProcessInfo::_name
Currently, there is an issue where 'exec [program]' results in tab
titles not being updated and not showing the process name as [program].
This is probably due to how 'exec' does not result in any change in
pid and simply replaces the current process image with a different one.

Due to a lack of pid changes, Session::updateSessionProcessInfo will not
call ProcessInfo::newInstance, thus not updating the ProcessInfo::_name.

This new change periodically checks and updates the _name of ProcessInfo
through ProcessInfo::readProcessName, which is called by
ProcessInfo::update.

BUG: 391128