Don't select half a CJK character
This was a visual regression.  Code using the selection already takes
care of dealing with CJK wide characters (mostly), and the old rendering
in TerminalPainter also did.  The new rendering is split in several
layers, which need to know about the selection to use the proper
background and foreground colors, but don't necessarily need to know
about characters, and this caused the regression.

To prevent trouble, we now take care of CJK wide characters before
rendering, when obtaining the characters from the model (Screen), where
characters are marked as selected if they fall into the selection
coordinates.

BUG: 474055