Provide compositor opaque hint
Konsole requests alpha channel even when background is opaque. Since the
opaque region is not set, the compositor (kwin) will render windows
beneath konsole because it has no idea what part of konsole is actually
opaque.

This change makes Konsole set the Qt::WA_TranslucentBackground attribute
based on whether opacity is set in the color scheme profile.

If the opacity is 1, the translucent background attribute will be unset
and QtWayland will set an opaque region, which would allow kwin to avoid
repainting windows behind konsole.

If the opacity changes to a value other than 1, the translucent
background flag will be set and QtWayland will unset the opaque region.