Show wallpaper on non-translucent top-levels
For reasons, showing the background image (wallpaper) has become
dependant on the ability of the top-level window to be translucent.

This does not need to be so.  Since 5bac30ab3, to draw the wallpaper,
first a background color is used to fill the dirty region, and then the
wallpaper is drawn, and the opacity of the background color and the
wallpaper are separate, so we can fill with a totally opaque background
color, and then draw a maybe translucent wallpaper over that, and no
glitches should appear.

BUG: 477800
(cherry picked from commit 8939910b61f2b9fd035a9bf9292dcdbda7ef17eb)