Implemented DBus methods for copying input to other sessions
Implemented methods which allow input to be copied to no other
session and to all sessions. In addition, users can also select
exactly which sessions to copy input to by passing a list of
Session ids. Calling supporting method copyingSessions() on a
session lists other sessions which are copying input from it.
Calling supporting method feederSessions() on a session lists
other sessions it is copying input from.

BUG: 307089