Implemented DBus methods for reading displayed text
The four methods allow users to retrieve the text displayed on a
view.

1. Session::getAllDisplayedText
2. Session::getAllDisplayedTextList
3. Session::getDisplayedText
4. Session::getDisplayedTextList

They can choose the return value to be a single string. Or call
the "...List" methods to get a list where every list element is
the text on a single line.

For methods "getAllDisplayedText" and "getAllDisplayedTextList",
the user can specify if trailing empty lines should be removed.

For the other two methods, the user can specify the range of
lines to be retrieved.

BUG: 238032