fix: Issue with focus setting on wdg after relocating to a new splitter
The bug is a problem where the widget was not receiving focus after
being moved to a new splitter (not clear why). The solution involves
connecting it to a signal, introducing a delay for the focus request.

The `deleteLater` call has been removed. This isn't necessary because
the old splitter after lose the remaining widget will subsequently
trigger the `childEvent` with `count == 0` and be deleted.

BUG: 479858