Fix the mouse-position base for semantic click
In the current master, the mouse click of the semantic shell
integration does not work on the top line in the terminal display.

This is caused by a discrepancy in the coordinates specified to
Vt102Emulation::sendMouseEvent.  Although the function
Vt102Emulation::sendMouseEvent expects 1-based coordinates (x,y) for
the mouse position, it tries to receive and process 0-based
coordinates only for the semantic mouse click.  However, the event is
rejected on the first line (y=0) and the first column (x=0) by the
bound check at the beginning of the function.

This patch fixes the discrepancy by switching to the 1-based
coordinates also for the semantic mouse click.