Support for OSC 99 (Kitty) desktop notifications
Konsole already supports the rxvt notification protocol [1] protocol
(OSC 777). For example, this command triggers a desktop notification
(provided this is enabled in the Konsole configuration):

  printf "\e]777;notify;%s;%s\e\\" title body

The desktop notification protocol implemented by Kitty is an
alternative protocol (based on [Free Desktop terminal WG Tcket #13 [2]
(OSC 99) for the same goal which can be tested using either of:

  printf \
    '\x1b]99;i=1:d=0;%s\x1b\\\x1b]99;i=1:d=1:p=body;%s\x1b\\' \
    title body
  printf \
    '\x1b]99;i=1:d=0:e=1;%s\x1b\\\x1b]99;i=1:d=1:p=body:e=1;%s\x1b\\' \
    "$(echo -n title | base64)" "$(echo -n body | base64)"

One benefit is that it supports arbitrary characters in the title and
body (by using base64-encoding).

Konsole already supports [3] some part of the Kitty graphics protocol
[4].

Note that there is another notification protocol from iTerm2 [5] (OSC
9):

printf "\e]9;%s\e\\" message

OSC 9 is not implemented in Konsole and not implemented in this PR.

[1] https://chromium.googlesource.com/apps/libapps/+/master/hterm/doc/ControlSequences.md#OSC
[2] https://gitlab.freedesktop.org/terminal-wg/specifications/-/issues/13
[3] https://invent.kde.org/utilities/konsole/-/commit/89101dc5b9c44c9f6f5f603f3b48db2aac6289cf
[4] https://sw.kovidgoyal.net/kitty/graphics-protocol/
[5] https://iterm2.com/documentation-escape-codes.html