Reshow configuration dialog after toggling borders
Usually, a QDialog with a parent stays on top of its parent.  However,
after hiding and showing the main window when toggling the
border-less/title-less option, this stops being true.

This can confuse users, if they apply this setting without closing the
configuration dialog, and the dialog gets behind the main window.
Afterwards, trying to open the configuration dialog doesn't work (it's
already open, but behind the window).

So, hide and show also the configuration dialog, which makes the dialog
stay on top of its parent again.

BUG: 479081