Add dummy title option for compatibility
On X11 Qt adds several commandline options automatically, including --title.

On Wayland that does not happen.

Some software using Konsole, like e.g. the VirtualBox Guest Additions installer,
expect that option to be there. Currently in that case Konsole will fail to start
because the option is unknown.

We already have -T for a similar compatibility problem, so register --title as
an alias to it