Add "No wrap" setting to search options
In some scenarios, users don't want to start a search again from the
top when they reach the end of the contents being searched. However,
if the contents are too large, it can be difficult to notice when
such jumping occurs. This commit aims to address this by adding an
option that disables the automatic wrapping in searches.

FEATURE: 303485