Supress incorrect resize notifications
There are paths that can lead to multiple resize events that end up back
at the same size.

An easy one to reproduce is a split view with the session toolbar
enabled but not the main toolbar. The session toolbar belongs to each
tab; when we switch view that is recreated by kxmlgui which results in
the toolbar deleting and creating causing two resize events on the
display.

This patch defers the resize notification until there is a paint at a
new size.

(cherry picked from commit 02a937cdc381f225072c6817722bc93976d00437)