Fixed the window geometry config file placement.
Since the window geometry placement was moved to the state config
files globally in KDE, exactly these files should be checked for
such configuration data (instead of the main application config file).
This issue may not been reproduced on the systems where the
'konsolerc' file remains filled by some old version of the konsole
so that it holds the window geometry. But on the fresh install, the
"Remember window size" option does nothing because the window geometry
is being stored in the '.local/share/konsole/konsolestaterc', but is
being looked up in the '.config/konsolerc' files.

CCBUG: 481898
CCBUG: 482954