Disable DBus (and X11) on Windows/macOS/Android per default
Like Kate, one should not try to use DBus on platform that not
support it. At the moment we have hacks in the Craft blueprints
to avoid the worst, but not using it is much nicer.

If the user wants to compile it, we have a USE_DBUS CMake option.