Add functionality for markers on scrollbar
Users can add markers to the scrollbar by double clicking on the
scrollbar. A user can remove a marker by double clicking on it again.

The position of the marker is stored internally as a percentage offset
of the scrollbar's length and is not attached to any line of output.
This means that on resize, a scrollbar marker originally created at
x percentile of the scrollbar's length from the top will still remain
at x percentile of the scrollbar's length from the top.

In the "Scrolling" tab in the settings, there are options for the
user to change the color of the markers and the size of the markers.
The size of the marker can be specified as a floating point number,
representing the percentage of the entire scrollbar's length it occupies.

GUI:
FEATURE: 338901