This shall fix a minor bug calling hosts that are defined in ~/.ssh/config by their profile name rather than its HostName.
    Definition:
    "profile name" is called Host in .ssh/config.

Example config in ~/.ssh/config:

 Host core
   HostName core.company.com
   User sshadmin

You want Konsole to run "ssh Host" or "ssh core", while
currently Konsole runs "ssh HostName" or "ssh core.company.com".
In the end it does not match any entity in ssh's config.

That error I think is laying in sshmanagerplugin.cpp:205-216
where we need to evolve the if statement.

Let me know in case something needs to be aligned with your requirements.
I am new to sw dev and this is my very first patch to a KDE or any cpp/Qt project.