fix calling hosts that are defined in ~/.ssh/config
This shall fix a minor bug calling hosts that are defined in ~/.ssh/config by their identifier name rather than its HostName.

> Definition:
> "identifier" is called Host in .ssh/config.

Example config in ~/.ssh/config:

> Host core
> 
> HostName core.company.com
> 
> User sshadmin

You want Konsole to run "ssh core", since calling ssh with a
config identifier makes ssh pull all config from .ssh/config.

Currently Konsole runs "ssh core.company.com" so it takes the hostname
instead of the config identifier.

In the end it does not match any entity in ssh's config and Konsole
would need to establish a correct set of parameters in the call to the
ssh binary.

That error I think is laying in sshmanagerplugin.cpp:205-216
where we need to evolve the if statement.

There was already a merge request for this issue report:

> BUG: https://bugs.kde.org/show_bug.cgi?id=450891

> MR: https://invent.kde.org/utilities/konsole/-/merge_requests/1055

But I think that MR is a bit of a shortcoming and does not take full
advantage of having the ssh binary do its job.

Let me know in case something needs to be aligned with your requirements.

I am new to sw dev and this is my very first patch to a KDE or any cpp/Qt project.