Implement scrollbar highlighting for search results
This PR introduces a feature to highlight the position of search results on the scrollbar, similar to how search results are highlighted in browsers like Google Chrome, editors like VSCode. The aim is to provide users with a visual indicator on the scrollbar, helping them quickly navigate through the results. 

FEATURE: 302284