'Allow escape sequences for links' setting fix in new sessions
The setting 'Allow escape sequences for links' is not applied to new
tabs and instances of Konsole. This is confusing when a user sets the
setting to enabled/allow. It only works after applying a profile, so
after changing the setting and applying the changes it works for the
current tab, but not for any new tabs or Konsole windows.

BUG: 469511
(cherry picked from commit ca51c19ad856ebdfa42b2d44b48d502ed81d3b0e)