Scale curly and double underlines with font size
Curly underlines are implemented by drawing two parallel, complementary-dashed
lines with a distance of 2 pixels.

This means they dont' scale with font size. They only look good with a small
font size. At a first glance, they are also not so easy to tell apart from
regular underlines.

Fix this by
1. drawing a cosine wave instead of of parallel dashed lines
2. always drawing a full wave period per glph, indendent of the font size
3. using the text line width, instead of always using 1px

A similar (but much less prominent) issue exists for double underlines.
Change them too, keeping the implementation consistent.

Tested with

    printf '\033[31m\033[4:3mcurly\033[4m\033[31m\033[4:2mdouble\033[4m'

Note that the curly underline overlaps with the "y".  Perhaps we should
fix this.

Probably there are more improvements to be made, and it's possible there is
a more elegant solution than repeated drawArc() calls.

Publishing this early nevertheless because it's hopefully close to an overall
improvement; But I did not have a chance to look into the bugs mentioned in
76f879cd7 (Draw characters in exact positions, 2022-08-06)