Fixes related to cursor artifacts
This commit addresses some artifacts related to cursor updates.

First issue is reported in #431171. It seems that an offset was
intentionally left out when refreshing the cursor background.

Second issue is related to resolved issue #483197: antialiasing
when drawing is not properly working as reported in QTBUG-66036.
Enabling antialiasing only for the outline should prevent
artifacts while still rendering a proper outline.

Other observed artifacts were fixed.

BUG: 483197
BUG: 431171