Inherit parent's app ID when creating a new cgroup for tabs
When creating a cgroup for a tab, we would use a hardcoded string as
cgroup name that included `org.kde.konsole` as app ID. This leads to
tabs incorrectly being attributed to Konsole in System Monitor's
application page, even if the tab is actually embedded in a different
application such as Kate.

Instead of hardcoding the app ID, use the application's desktop file
name, so that if we lookup the app ID from a cgroup name, we get back to
the original application instead of Konsole.

BUG: 508716