Fix regression: Unix locale encoding is UTF-8
On c7e02ed8 we moved from the deprecated QTextCodec to QStringDecoder
and QStringEncoder.  According to Qt's QStringConverter documentation
and source code, the system ("Locale") encoding in Unix based systems is
assumed to be UTF-8.  So, we should return true on Emulation::utf8()
when the current encoding is the system encoding.

Failure to do so was causing the pty to not have the IUTF8 mode set,
causing problems when backspacing as reported on
https://discuss.kde.org/t/konsole-backspace-utf-8/35608/2
where the reporter Mikhail_Ramendik writes:

  When I use Konsole and press backspace when entering UTF-8 non-ASCII
  text (Cyrillic in my case), weird characters appear in the input that
  the program I run gets. They cause Python programs to crash, being
  unable to process the encoding. If I just do “cat > qqq” and use
  backspace, I get characters like this:

  [mramendi@ramendik-2021 ~]$ cat > qqq
  пеф [backspace was used before the last letter was entered]
  [mramendi@ramendik-2021 ~]$ cat qqq
  пе�ф