Add two keyboard actions for focusing on next/prev in split view mode
This commit adds two keyboard action for focusing on next or previous
view in split view tab.

The default shortcuts are `ctrl-F11` for next view and `ctrl=shift-F11`
for previous view. I find these shortcuts more useful in some situations
than the currently available shortcuts for moving the focus
up/down/left/right. The order of the views in the list is the one
returned by `ViewSplitter::findChildren<TerminalDisplay *>`, so it
does not correlate exactly to their positions.