Fix context menu event handling
Konsole was using mousePressEvent with button == Qt::RightButton for
emulating a context menu event, which is wrong because there is a
dedicated event for context menu.

Due to this there were weird issues on wayland for e.g., you open the
context menu and then click somewhere else -> you get another context menu
if the target widget supports one. This happened because after
mousePressEvent, there is a mouseReleaseEvent, which konsole doesn't
handle always which then gets synthesized into a context menu event
and delivered to the widget where the click happened.

This is also a bug in Qt wayland I think and should be fixed there but
there is no reason for Konsole not to handle this properly.

To fix, move the logic for context menu event to contextMenuEvent.

BUG: 512130