Fix spurious home directory replacement in window title
If I run

	echo -e "\033]7;file:///home/$(whoami)SUFFIX\033\\"

then the default window title will include "~SUFFIX", which makes no
sense because SUFFIX is not a file inside "~".  It should rather be
something like "/home/johannesSUFFIX".

Make it so.

I'm not sure if we can rely on path separator being "/" (on Windows)
but it does look like we rely on this elsewhere in our code.

BUG: 513684